Part 1: packet handlers, move to separate classes
